Refactor CI jobs and client-java build & tests
